+++ Compiler/Compiler/scanner/keyword.swift
@@ -9,8 +9,8 @@ class KeywordProvider {
 			// relevant for <Identifier>
 			"bool": Token(terminal: Terminal.TYPE,
 				attribute: Token.Attribute.Type(Token.TypeIdentifier.BOOLEAN)),
-			"int64": Token(terminal: Terminal.TYPE,
-				attribute: Token.Attribute.Type(Token.TypeIdentifier.INT64)),
+			"int32": Token(terminal: Terminal.TYPE,
+				attribute: Token.Attribute.Type(Token.TypeIdentifier.INT32)),
 			"call": Token(terminal: Terminal.CALL),
 			"do": Token(terminal: Terminal.DO),
 			"else": Token(terminal: Terminal.ELSE),
